рабочее приложение
