--- MovieQuiz/Models/QuizQuestions.swift
@@ -1,8 +1,0 @@
-
-import Foundation
-
-struct QuizQuestion {
-    let image: String
-    let text: String
-    let correctAnswer: Bool
-}
