--- MovieQuiz/Models/QuizResultsViewModel.swift
@@ -1,8 +1,0 @@
-
-import Foundation
-
-struct QuizResultsViewModel {
-    let title: String
-    let text: String
-    let buttonText: String
-}
