--- MovieQuiz/Models/QuizStepViewModel.swift
@@ -1,8 +1,0 @@
-
-import UIKit
-
-struct QuizStepViewModel {
-    let image: UIImage
-    let question: String
-    let questionNumber: String
-}
